move function to service file; add replaceMissingValues to InversionSolutionMfdTab;
